add return values checks for commands

+++ fortscale/fortscale-webapp/wardeploy.sh
@@ -10,8 +10,27 @@ dst_dir="/var/lib/tomcat6/webapps/fortscale-webapp"
 dst_file="${dst_dir}.war"
 
 # cleanup old files
-sudo /bin/rm -fr "${dst_dir}"
+if [ -d ${dst_dir} ];
+    sudo /bin/rm -fr "${dst_dir}"
+    res=$?
+    if [ $res -ne 0 ]; then
+        echo "FATAL: Failed to clean up old webapp folder, exiting"
+        exit $res
+    fi
+fi
 # copy new war
 cp -pr ${src_file} ${dst_file}
+res=$?
+if [ $res -ne 0 ]; then
+    echo "FATAL: Failed to deploy the new war file, exiting"
+    exit $res
+fi
+
 # restart service
 sudo /sbin/service tomcat6 restart
+res=$?
+if [ $res -ne 0 ]; then
+    echo "FATAL: Failed to restart tomcat6 service, exiting"
+    exit $res
+fi
+exit 0
